pyproject.toml
[project]
name = "status-light"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "platformio>=6.1.19",
    "pyserial>=3.5",
]
